# backend/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  auth.Session:
    properties:
      created_at:
        type: string
      expire_at:
        type: string
      id:
        type: integer
      user_agent:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_auth.SessionsResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/auth.Session'
        type: array
    type: object
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  models.AgentAddressType:
    enum:
    - ip
    - domain
    type: string
    x-enum-varnames:
    - AgentAddressTypeIP
    - AgentAddressTypeDomain
...
  models.OcservAgent:
    properties:
      address:
        maxLength: 255
        type: string
      address_type:
        allOf:
        - $ref: '#/definitions/models.AgentAddressType'
        enum:
        - ip
        - domain
      created_at:
        type: string
      id:
        type: integer
      name:
        maxLength: 128
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      token:
        maxLength: 512
        type: string
      updated_at:
        type: string
    required:
    - address
    - address_type
    - name
    - token
    type: object
...
  models.OcservInfo:
    properties:
      status:
        type: string
      version:
        $ref: '#/definitions/models.ServerVersion'
    required:
    - status
    - version
    type: object
...
  models.ServerVersion:
    properties:
      occtl_version:
        type: string
      ocserv_version:
        type: string
    type: object
...
  ocserv_agent.CreateInput:
    properties:
      address:
        maxLength: 255
        type: string
      address_type:
        allOf:
        - $ref: '#/definitions/models.AgentAddressType'
        enum:
        - ip
        - domain
      name:
        maxLength: 128
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      token:
        maxLength: 512
        type: string
    required:
    - address
    - address_type
    - name
    - token
    type: object
  ocserv_agent.UpdateInput:
    properties:
      address:
        maxLength: 255
        type: string
      address_type:
        allOf:
        - $ref: '#/definitions/models.AgentAddressType'
        enum:
        - ip
        - domain
      name:
        maxLength: 128
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      token:
        maxLength: 512
        type: string
    required:
    - address
    - address_type
    - name
    - token
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
...
info:
  contact: {}
  description: Administrative API for managing Ocserv Dashboard users, groups, reports,
    runtime, backups, system settings, Telegram integrations, and agent nodes.
  title: Ocserv Dashboard Admin API
  version: "1.0"
paths:
  /auth/logout:
    post:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Logout current session
      tags:
      - Auth
  /auth/sessions:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_auth.SessionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: List active user sessions
      tags:
      - Auth
  /auth/sessions/{id}:
    delete:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Revoke user session
      tags:
      - Auth
...
  /occtl/commands:
    get:
      consumes:
      - application/json
      description: Occtl Commands
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Command Action ID (1 to 15)
        in: query
        name: action
        required: true
        type: integer
      - description: Optional parameter depending on command
        in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Occtl Commands
      tags:
      - OCCTL
  /occtl/server_info:
    get:
      consumes:
      - application/json
      description: Server information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OcservInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Server information
      tags:
      - OCCTL
  /ocserv/agents:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OcservAgent'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: List Ocserv agents
      tags:
      - Ocserv Agents
    post:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Agent
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_agent.CreateInput'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservAgent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Create Ocserv agent
      tags:
      - Ocserv Agents
  /ocserv/agents/{id}:
    delete:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Delete Ocserv agent
      tags:
      - Ocserv Agents
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OcservAgent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Get Ocserv agent
      tags:
      - Ocserv Agents
    patch:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Agent ID
        in: path
        name: id
        required: true
        type: integer
      - description: Agent
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_agent.UpdateInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OcservAgent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Update Ocserv agent
      tags:
      - Ocserv Agents
...
produces:
- application/json
swagger: "2.0"